+++ src/plans/objective.jl
@@ -6,7 +6,7 @@ An abstract type to specify the kind of evaluation a [`AbstractManifoldObjective
 abstract type AbstractEvaluationType end
 
 @doc raw"""
-    AbstractManifoldObjective{T<:AbstractEvaluationType}
+    AbstractManifoldObjective{E<:AbstractEvaluationType}
 
 Describe the collection of the optimization function ``f\colon \mathcal M → \bbR` (or even a vectorial range)
 and its corresponding elements, which might for example be a gradient or (one or more) prxomial maps.
@@ -20,7 +20,16 @@ All these elements should usually be implemented as functions
 
 the type `T` indicates the global [`AbstractEvaluationType`](@ref).
 """
-abstract type AbstractManifoldObjective{T<:AbstractEvaluationType} end
+abstract type AbstractManifoldObjective{E<:AbstractEvaluationType} end
+
+@doc raw"""
+    AbstractDecoratedManifoldObjective{E<:AbstractEvaluationType,O<:AbstractManifoldObjective}
+
+A common supertype for all decorators of [`AbstractManifoldObjective`](@ref)s to simplify dispatch.
+    The second parameter should refer to the undecorated objective (i.e. the most inner one).
+"""
+abstract type AbstractDecoratedManifoldObjective{E,O<:AbstractManifoldObjective} <:
+              AbstractManifoldObjective{E} end
 
 @doc raw"""
     AllocatingEvaluation <: AbstractEvaluationType
@@ -38,6 +47,25 @@ do not allocate memory but work on their input, i.e. in place.
 """
 struct InplaceEvaluation <: AbstractEvaluationType end
 
+struct ReturnManifoldObjective{E,P,O<:AbstractManifoldObjective{E}} <:
+       AbstractDecoratedManifoldObjective{E,P}
+    objective::O
+end
+function ReturnManifoldObjective(
+    o::O
+) where {E<:AbstractEvaluationType,O<:AbstractManifoldObjective{E}}
+    return ReturnManifoldObjective{E,O,O}(o)
+end
+function ReturnManifoldObjective(
+    o::O
+) where {
+    E<:AbstractEvaluationType,
+    P<:AbstractManifoldObjective,
+    O<:AbstractDecoratedManifoldObjective{E,P},
+}
+    return ReturnManifoldObjective{E,P,O}(o)
+end
+
 """
     dispatch_objective_decorator(o::AbstractManoptSolverState)
 
@@ -49,6 +77,7 @@ Decorators indicate this by returning `Val{true}` for further dispatch.
 The default is `Val{false}`, i.e. by default an state is not decorated.
 """
 dispatch_objective_decorator(::AbstractManifoldObjective) = Val(false)
+dispatch_objective_decorator(::AbstractDecoratedManifoldObjective) = Val(true)
 
 """
     is_object_decorator(s::AbstractManifoldObjective)
@@ -60,15 +89,48 @@ function is_objective_decorator(s::AbstractManifoldObjective)
 end
 
 @doc raw"""
-    get_objective(o::AbstractManifoldObjective)
+    get_objective(o::AbstractManifoldObjective, recursive=true)
 
-return the undecorated [`AbstractManifoldObjective`](@ref) of the (possibly) decorated `o`.
+return the (one step) undecorated [`AbstractManifoldObjective`](@ref) of the (possibly) decorated `o`.
 As long as your decorated objective stores the objetive within `o.objective` and
 the [`dispatch_objective_decorator`](@ref) is set to `Val{true}`,
 the internal state are extracted automatically.
+
+By default the objective that is stored within a decorated objective is assumed to be at
+`o.objective`. Overwrtie `_get_objective(o, ::Val{true}, recursive) to change this bevahiour for your objective `o`
+for both the recursive and the nonrecursive case.
+
+If `recursive` is set to `false`, only the most outer decorator is taken away instead of all.
 """
-function get_objective(o::AbstractManifoldObjective)
-    return _get_objective(o, dispatch_objective_decorator(o))
+function get_objective(o::AbstractManifoldObjective, recursive=true)
+    return _get_objective(o, dispatch_objective_decorator(o), recursive)
 end
-_get_objective(o::AbstractManifoldObjective, ::Val{false}) = o
-_get_objective(o::AbstractManifoldObjective, ::Val{true}) = get_objective(o.objective)
+_get_objective(o::AbstractManifoldObjective, ::Val{false}, rec) = o
+function _get_objective(o::AbstractManifoldObjective, ::Val{true}, rec)
+    return rec ? get_objective(o.objective) : o.objective
+end
+function status_summary(o::AbstractManifoldObjective{E}) where {E}
+    return "$(nameof(typeof(o))){$E}"
+end
+# Default undecorate for summary
+function status_summary(co::AbstractDecoratedManifoldObjective)
+    return status_summary(get_objective(co, false))
+end
+
+function show(io::IO, o::AbstractManifoldObjective{E}) where {E}
+    return print(io, "$(nameof(typeof(o))){$E}")
+end
+# Default undecorate for show
+function show(io::IO, co::AbstractDecoratedManifoldObjective)
+    return show(io, get_objective(co, false))
+end
+
+function show(io::IO, t::Tuple{<:AbstractManifoldObjective,P}) where {P}
+    return print(
+        io,
+        """
+$(status_summary(t[1]))
+
+To access the solver result, call `get_solver_result` on this variable.""",
+    )
+end