+++ src/solvers/augmented_Lagrangian_method.jl
@@ -261,8 +261,8 @@ function augmented_Lagrangian_method(
     return augmented_Lagrangian_method!(M, cmo, q; evaluation=evaluation, kwargs...)
 end
 function augmented_Lagrangian_method(
-    M::AbstractManifold, cmo::ConstrainedManifoldObjective, p=rand(M); kwargs...
-)
+    M::AbstractManifold, cmo::O, p=rand(M); kwargs...
+) where {O<:Union{ConstrainedManifoldObjective,AbstractDecoratedManifoldObjective}}
     q = copy(M, p)
     return augmented_Lagrangian_method!(M, cmo, q; kwargs...)
 end
@@ -319,7 +319,7 @@ function augmented_Lagrangian_method!(
 end
 function augmented_Lagrangian_method!(
     M::AbstractManifold,
-    cmo::ConstrainedManifoldObjective,
+    cmo::O,
     p;
     evaluation=AllocatingEvaluation(),
     ϵ::Real=1e-3,
@@ -360,7 +360,7 @@ function augmented_Lagrangian_method!(
         StopWhenSmallerOrEqual(:ϵ, ϵ_min) & StopWhenChangeLess(1e-10)
     ),
     kwargs...,
-)
+) where {O<:Union{ConstrainedManifoldObjective,AbstractDecoratedManifoldObjective}}
     alms = AugmentedLagrangianMethodState(
         M,
         cmo,
@@ -383,7 +383,8 @@ function augmented_Lagrangian_method!(
     dcmo = decorate_objective!(M, cmo; kwargs...)
     mp = DefaultManoptProblem(M, dcmo)
     alms = decorate_state!(alms; kwargs...)
-    return get_solver_return(solve!(mp, alms))
+    solve!(mp, alms)
+    return get_solver_return(get_objective(mp), alms)
 end
 
 #
@@ -410,14 +411,14 @@ function step_solver!(mp::AbstractManoptProblem, alms::AugmentedLagrangianMethod
 
     # update multipliers
     cost_ineq = get_inequality_constraints(mp, alms.p)
-    n_ineq_constraint = size(cost_ineq, 1)
+    n_ineq_constraint = length(cost_ineq)
     alms.μ .=
         min.(
             ones(n_ineq_constraint) .* alms.μ_max,
             max.(alms.μ .+ alms.ρ .* cost_ineq, zeros(n_ineq_constraint)),
         )
     cost_eq = get_equality_constraints(mp, alms.p)
-    n_eq_constraint = size(cost_eq, 1)
+    n_eq_constraint = length(cost_eq)
     alms.λ =
         min.(
             ones(n_eq_constraint) .* alms.λ_max,