+++ src/solvers/solver.jl
@@ -64,9 +64,11 @@ decorate the [`AbstractManifoldObjective`](@ref)` o` with specific decorators.
 optional arguments provide necessary details on the decorators.
 A specific one is used to activate certain decorators.
 
-* `cache` – (`missing`) currently only supports the [`SimpleCacheObjective`](@ref)
-  which is activated by either specifying the symbol `:Simple` or the tuple
-  (`:Simple, kwargs...`) to pass down keyword arguments
+* `cache` – (`missing`) specify a cache. Currenlty `:Simple` is supported and `:LRU` if you
+  load `LRUCache.jl`. For this case a tuple specifying what to cache and how many can be provided,
+  i.e. `(:LRU, [:Cost, :Gradient], 10)`, where the number specifies the size of each cache.
+  and 10 is the default if one omits the last tuple entry
+* `count` – (`missing`) specify calls to the objective to be called, see [`ManifoldCountObjective`](@ref) for the full list
 
 other keywords are ignored.
 
@@ -75,9 +77,20 @@ other keywords are ignored.
 [`objective_cache_factory`](@ref)
 """
 function decorate_objective!(
-    M::AbstractManifold, o::O; cache::Union{Missing,Symbol}=missing, kwargs...
-) where {O<:AbstractManifoldObjective}
-    deco_o = ismissing(cache) ? o : objective_cache_factory(M, o, cache)
+    M::AbstractManifold,
+    o::O;
+    cache::Union{
+        Missing,Symbol,Tuple{Symbol,<:AbstractArray},Tuple{Symbol,<:AbstractArray,P}
+    }=missing,
+    count::Union{Missing,AbstractVector{<:Symbol}}=missing,
+    return_objective=false,
+    kwargs...,
+) where {O<:AbstractManifoldObjective,P}
+    # Order: First count _then_ cache, so that cache is the outer decorator and
+    # we only count _after_ cache misses
+    deco_o = ismissing(count) ? o : objective_count_factory(M, o, count)
+    deco_o = ismissing(cache) ? deco_o : objective_cache_factory(M, deco_o, cache)
+    deco_o = return_objective ? ReturnManifoldObjective(deco_o) : deco_o
     return deco_o
 end